Update integer decode test to use max & min values
This commit updates the decode tests for type "integer" to decode the
maximum (by absolute value) allowed values. Previously, the test
attempted to decode +/- 5_000_000, which are well within the bounds of
the decoder's limits for type integer. The new values reflect the limits
of the decoder, allowing the tests to serve as documentation.

Notes:

- The Avro spec says that type integer can handle 32-bit signed
integers, but the number above is the maximum 28-bit signed integer.
- Lines 42, 43, 48, and 49 of decode_test.exs are meant to test type
long (64-bit signed integer) but are actually testing the max/min 32-bit
signed integer.